# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type PersonPlanStatus = typeof PersonPlanStatus[keyof typeof PersonPlanStatus];


export const PersonPlanStatus = {
  'on-track': 'on-track',
  'needs-attention': 'needs-attention',
  'getting-started': 'getting-started',
} as const;

export interface Person {
  id: number;
  name: string;
  relationship: string;
  age: number;
  planStatus: PersonPlanStatus;
  initials: string;
  accent: string;
}

export interface PersonInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  relationship: string;
  /** @minimum 0 */
  age: number;
}

export type PersonUpdatePlanStatus = typeof PersonUpdatePlanStatus[keyof typeof PersonUpdatePlanStatus];


export const PersonUpdatePlanStatus = {
  'on-track': 'on-track',
  'needs-attention': 'needs-attention',
  'getting-started': 'getting-started',
} as const;

export interface PersonUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  relationship?: string;
  /** @minimum 0 */
  age?: number;
  planStatus?: PersonUpdatePlanStatus;
}

export type TaskStatus = typeof TaskStatus[keyof typeof TaskStatus];


export const TaskStatus = {
  todo: 'todo',
  'in-progress': 'in-progress',
  done: 'done',
} as const;

export type TaskPriority = typeof TaskPriority[keyof typeof TaskPriority];


export const TaskPriority = {
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface Task {
  id: number;
  title: string;
  category: string;
  dueLabel: string;
  status: TaskStatus;
  priority: TaskPriority;
}

export type TaskInputPriority = typeof TaskInputPriority[keyof typeof TaskInputPriority];


export const TaskInputPriority = {
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface TaskInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  category: string;
  dueLabel?: string;
  priority?: TaskInputPriority;
}

export type TaskUpdateStatus = typeof TaskUpdateStatus[keyof typeof TaskUpdateStatus];


export const TaskUpdateStatus = {
  todo: 'todo',
  'in-progress': 'in-progress',
  done: 'done',
} as const;

export type TaskUpdatePriority = typeof TaskUpdatePriority[keyof typeof TaskUpdatePriority];


export const TaskUpdatePriority = {
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface TaskUpdate {
  /** @minLength 1 */
  title?: string;
  status?: TaskUpdateStatus;
  priority?: TaskUpdatePriority;
  dueLabel?: string;
}

export type DocumentStatus = typeof DocumentStatus[keyof typeof DocumentStatus];


export const DocumentStatus = {
  ready: 'ready',
  review: 'review',
  missing: 'missing',
} as const;

export interface Document {
  id: number;
  name: string;
  type: string;
  updatedLabel: string;
  status: DocumentStatus;
}

export type BenefitStatus = typeof BenefitStatus[keyof typeof BenefitStatus];


export const BenefitStatus = {
  protected: 'protected',
  review: 'review',
  incomplete: 'incomplete',
} as const;

export interface Benefit {
  name: string;
  status: BenefitStatus;
  note: string;
}

export type ActivityTone = typeof ActivityTone[keyof typeof ActivityTone];


export const ActivityTone = {
  green: 'green',
  blue: 'blue',
  amber: 'amber',
} as const;

export interface Activity {
  id: number;
  label: string;
  timeLabel: string;
  tone: ActivityTone;
}

export interface Dashboard {
  person: Person;
  /**
     * @minimum 0
     * @maximum 100
     */
  progress: number;
  completedTasks: number;
  totalTasks: number;
  benefits: Benefit[];
  netWorth: number;
  monthlySupport: number;
  recentActivity: Activity[];
}