# lib/api-zod/src/generated/types/activity.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityTone } from './activityTone';

export interface Activity {
  id: number;
  label: string;
  timeLabel: string;
  tone: ActivityTone;
}