# lib/api-zod/src/generated/types/benefit.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BenefitStatus } from './benefitStatus';

export interface Benefit {
  name: string;
  status: BenefitStatus;
  note: string;
}