# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Activity } from './activity';
import type { Benefit } from './benefit';
import type { Person } from './person';

export interface Dashboard {
  person: Person;
  /**
     * @minimum 0
     * @maximum 100
     */
  progress: number;
  completedTasks: number;
  totalTasks: number;
  benefits: Benefit[];
  netWorth: number;
  monthlySupport: number;
  recentActivity: Activity[];
}