# lib/api-zod/src/generated/types/person.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PersonPlanStatus } from './personPlanStatus';

export interface Person {
  id: number;
  name: string;
  relationship: string;
  age: number;
  planStatus: PersonPlanStatus;
  initials: string;
  accent: string;
}