# lib/api-zod/src/generated/types/personInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PersonInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  relationship: string;
  /** @minimum 0 */
  age: number;
}