# lib/api-zod/src/generated/types/personUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PersonUpdatePlanStatus } from './personUpdatePlanStatus';

export interface PersonUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  relationship?: string;
  /** @minimum 0 */
  age?: number;
  planStatus?: PersonUpdatePlanStatus;
}