# lib/api-zod/src/generated/types/task.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TaskPriority } from './taskPriority';
import type { TaskStatus } from './taskStatus';

export interface Task {
  id: number;
  title: string;
  category: string;
  dueLabel: string;
  status: TaskStatus;
  priority: TaskPriority;
}