# lib/api-zod/src/generated/types/taskInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TaskInputPriority } from './taskInputPriority';

export interface TaskInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  category: string;
  dueLabel?: string;
  priority?: TaskInputPriority;
}