# lib/api-zod/src/generated/types/taskUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TaskUpdatePriority } from './taskUpdatePriority';
import type { TaskUpdateStatus } from './taskUpdateStatus';

export interface TaskUpdate {
  /** @minLength 1 */
  title?: string;
  status?: TaskUpdateStatus;
  priority?: TaskUpdatePriority;
  dueLabel?: string;
}